refactor: extract security component beans and refactor auth rules
