+++ src/test/kotlin/com/example/workflow/integration/infra/security/config/SecurityConfigIntegrationTest.kt
@@ -1,5 +1,6 @@
 package com.example.workflow.integration.infra.security.config
 
+import com.example.workflow.infra.security.config.SecurityComponentsConfig
 import com.example.workflow.infra.security.config.SecurityConfig
 import com.example.workflow.integration.test.controller.SecurityConfigTestController
 import com.example.workflow.integration.test.path.AuthenticatedPathsProvider
@@ -28,7 +29,7 @@ import org.springframework.test.web.servlet.assertj.MvcTestResult
 
 @IntegrationTest
 @WebMvcTest(SecurityConfigTestController::class)
-@Import(SecurityConfig::class, SecurityConfigIntegrationTest.MockConfig::class)
+@Import(SecurityConfig::class, SecurityComponentsConfig::class, SecurityConfigIntegrationTest.MockConfig::class)
 @ActiveProfiles("test", "security-test-controller")
 class SecurityConfigIntegrationTest {
     @Autowired
